requirements.txt
amqp==5.0.9
async-generator==1.10
attrs==21.4.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
configparser==5.2.0
crayons==0.4.0
cryptography==36.0.1
Deprecated==1.2.13
distro==1.6.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
flower==1.0.0
h11==0.13.0
humanize==3.13.1
idna==3.3
jsonschema==3.2.0
kombu==5.2.3
numpy==1.22.1
outcome==1.1.0
packaging==21.3
pandas==1.4.0
paramiko==2.9.2
prometheus-client==0.12.0
prompt-toolkit==3.0.24
pycparser==2.21
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==5.4.1
redis==4.1.1
requests==2.27.1
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
texttable==1.6.4
tornado==6.1
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
webdriver-manager==3.5.2
websocket-client==0.59.0
wrapt==1.13.3
wsproto==1.0.0
xlrd==2.0.1
XlsxWriter==3.0.2
